+++ src/app/loadout-drawer/LoadoutDrawerContents.tsx
@@ -1,10 +1,12 @@
 import { t } from 'app/i18next-t';
+import { SocketOverrides, SocketOverridesForItems } from 'app/inventory/store/override-sockets';
 import { getCurrentStore } from 'app/inventory/stores-helpers';
 import { itemCanBeInLoadout } from 'app/utils/item-utils';
 import { infoLog } from 'app/utils/log';
 import { DestinyClass } from 'bungie-api-ts/destiny2';
+import { BucketHashes } from 'data/d2/generated-enums';
 import _ from 'lodash';
-import React from 'react';
+import React, { useMemo } from 'react';
 import type {
   DimBucketType,
   InventoryBucket,
@@ -17,9 +19,9 @@ import { addIcon, AppIcon } from '../shell/icons';
 import { Loadout } from './loadout-types';
 import LoadoutDrawerBucket from './LoadoutDrawerBucket';
 import SavedMods from './SavedMods';
+import { Subclass } from './subclass-drawer/Subclass';
 
 const loadoutTypes: DimBucketType[] = [
-  'Class',
   'Primary',
   'Special',
   'Heavy',
@@ -78,6 +80,7 @@ export default function LoadoutDrawerContents(
     add,
     onOpenModPicker,
     removeModByHash,
+    onApplySocketOverrides,
   }: {
     loadout: Loadout;
     savedMods: PluggableInventoryItemDefinition[];
@@ -89,13 +92,14 @@ export default function LoadoutDrawerContents(
     add(item: DimItem, e?: MouseEvent, equip?: boolean): void;
     onOpenModPicker(): void;
     removeModByHash(itemHash: number): void;
+    onApplySocketOverrides(item: DimItem, socketOverrides: SocketOverrides): void;
   }
 ) {
   const itemsByBucket = _.groupBy(items, (i) => i.bucket.hash);
 
   function doFillLoadoutFromEquipped(e: React.MouseEvent) {
     e.preventDefault();
-    fillLoadoutFromEquipped(loadout, itemsByBucket, stores, add);
+    fillLoadoutFromEquipped(loadout, itemsByBucket, stores, add, onApplySocketOverrides);
   }
   function doFillLoadOutFromUnequipped(e: React.MouseEvent) {
     e.preventDefault();
@@ -108,23 +112,52 @@ export default function LoadoutDrawerContents(
     availableTypes,
     (bucket) => bucket.hash && itemsByBucket[bucket.hash] && itemsByBucket[bucket.hash].length
   );
+  const subclassBucket = buckets.byType.Class;
 
   const showFillFromEquipped = typesWithoutItems.some((b) => fromEquippedTypes.includes(b.type!));
 
+  const { subclassItems, subclassSocketOverrides } = useMemo(() => {
+    const subclassSocketOverrides: SocketOverridesForItems = {};
+    const subclassBucket = buckets.byType.Class;
+    const subclassItems: DimItem[] =
+      (subclassBucket?.hash && itemsByBucket[subclassBucket.hash]) || [];
+
+    for (const item of loadout.items) {
+      if (subclassItems.some((subclass) => subclass.id === item.id)) {
+        subclassSocketOverrides[item.id] = item.socketOverrides || {};
+      }
+    }
+    return { subclassSocketOverrides, subclassBucket, subclassItems };
+  }, [buckets.byType.Class, itemsByBucket, loadout.items]);
+
+  const showSubclassButton =
+    !loadout ||
+    loadout.classType === DestinyClass.Unknown ||
+    !subclassItems.length ||
+    subclassItems.every((i) => i.classType !== loadout.classType);
+
   return (
     <>
-      {typesWithoutItems.length > 0 && (
-        <div className="loadout-add-types">
-          {showFillFromEquipped && (
-            <a className="dim-button loadout-add" onClick={doFillLoadoutFromEquipped}>
-              <AppIcon icon={addIcon} /> {t('Loadouts.AddEquippedItems')}
-            </a>
-          )}
-          <a className="dim-button loadout-add" onClick={doFillLoadOutFromUnequipped}>
-            <AppIcon icon={addIcon} /> {t('Loadouts.AddUnequippedItems')}
+      <div className="loadout-add-types">
+        {showFillFromEquipped && (
+          <a className="dim-button loadout-add" onClick={doFillLoadoutFromEquipped}>
+            <AppIcon icon={addIcon} /> {t('Loadouts.AddEquippedItems')}
           </a>
-
-          {typesWithoutItems.map((bucket) => (
+        )}
+        <a className="dim-button loadout-add" onClick={doFillLoadOutFromUnequipped}>
+          <AppIcon icon={addIcon} /> {t('Loadouts.AddUnequippedItems')}
+        </a>
+        {showSubclassButton && (
+          <a
+            key={subclassBucket.type}
+            onClick={() => pickLoadoutSubclass(loadout, subclassBucket, subclassItems, add)}
+            className="dim-button loadout-add"
+          >
+            <AppIcon icon={addIcon} /> {subclassBucket.name}
+          </a>
+        )}
+        {typesWithoutItems.length > 0 &&
+          typesWithoutItems.map((bucket) => (
             <a
               key={bucket.type}
               onClick={() => pickLoadoutItem(loadout, bucket, add)}
@@ -133,24 +166,36 @@ export default function LoadoutDrawerContents(
               <AppIcon icon={addIcon} /> {bucket.name}
             </a>
           ))}
-          <a onClick={() => onOpenModPicker()} className="dim-button loadout-add">
-            <AppIcon icon={addIcon} /> {t('Loadouts.ArmorMods')}
-          </a>
-        </div>
-      )}
+        <a onClick={() => onOpenModPicker()} className="dim-button loadout-add">
+          <AppIcon icon={addIcon} /> {t('Loadouts.ArmorMods')}
+        </a>
+      </div>
       <div className="loadout-added-items">
-        {typesWithItems.map((bucket) => (
-          <LoadoutDrawerBucket
-            key={bucket.type}
-            bucket={bucket}
-            loadoutItems={loadout.items}
-            items={itemsByBucket[bucket.hash] || []}
-            pickLoadoutItem={(bucket) => pickLoadoutItem(loadout, bucket, add)}
+        {typesWithItems.map((bucket) =>
+          $featureFlags.loadoutSubclasses && bucket.type === 'Class' ? null : (
+            <LoadoutDrawerBucket
+              key={bucket.type}
+              bucket={bucket}
+              loadoutItems={loadout.items}
+              items={itemsByBucket[bucket.hash] || []}
+              pickLoadoutItem={(bucket) => pickLoadoutItem(loadout, bucket, add)}
+              equip={equip}
+              remove={remove}
+            />
+          )
+        )}
+      </div>
+      {subclassItems.length > 0 &&
+        subclassItems.map((subclass) => (
+          <Subclass
+            key={subclass.hash}
+            subclass={subclass}
+            socketOverrides={subclassSocketOverrides[subclass.id]}
             equip={equip}
             remove={remove}
+            onApplySocketOverrides={onApplySocketOverrides}
           />
         ))}
-      </div>
       <SavedMods
         savedMods={savedMods}
         onOpenModPicker={onOpenModPicker}
@@ -191,11 +236,71 @@ async function pickLoadoutItem(
   } catch (e) {}
 }
 
+async function pickLoadoutSubclass(
+  loadout: Loadout,
+  subclassBucket: InventoryBucket,
+  savedSubclasses: DimItem[],
+  add: (item: DimItem, e?: MouseEvent) => void
+) {
+  const loadoutClassType = loadout?.classType;
+  function loadoutHasItem(item: DimItem) {
+    return loadout?.items.some((i) => i.id === item.id && i.hash === item.hash);
+  }
+  function loadoutHasSubclassForClass(item: DimItem) {
+    return savedSubclasses.some(
+      (s) => item.bucket.type === 'Class' && s.classType === item.classType
+    );
+  }
+
+  try {
+    const { item } = await showItemPicker({
+      filterItems: (item: DimItem) =>
+        item.bucket.type === 'Class' &&
+        (!loadout ||
+          loadout.classType === DestinyClass.Unknown ||
+          item.classType === loadoutClassType) &&
+        itemCanBeInLoadout(item) &&
+        !loadoutHasSubclassForClass(item) &&
+        !loadoutHasItem(item),
+      // We can only sort so that the classes are grouped and stasis comes first
+      sortBy: (item) => `${item.classType}-${item.energy?.energyType}`,
+      prompt: t('Loadouts.ChooseItem', { name: subclassBucket.name }),
+
+      // don't show information related to selected perks so we don't give the impression
+      // that we will update perk selections when applying the loadout
+      ignoreSelectedPerks: true,
+    });
+
+    add(item);
+  } catch (e) {}
+}
+
+function createSocketOverridesFromEquipped(
+  item: DimItem,
+  onApplySocketOverrides: (item: DimItem, socketOverrides: SocketOverrides) => void
+) {
+  const socketOverrides: SocketOverrides = {};
+  for (const socket of item.sockets?.allSockets || []) {
+    // If the socket is plugged and we plug isn't the initial plug we apply the overrides
+    // to the loadout.
+    if (
+      socket.plugged &&
+      socket.plugged.plugDef.hash !== socket.socketDefinition.singleInitialItemHash
+    ) {
+      socketOverrides[socket.socketIndex] = socket.plugged.plugDef.hash;
+    }
+  }
+  if (Object.keys(socketOverrides).length) {
+    onApplySocketOverrides(item, socketOverrides);
+  }
+}
+
 function fillLoadoutFromEquipped(
   loadout: Loadout,
   itemsByBucket: { [bucketId: string]: DimItem[] },
   stores: DimStore[],
-  add: (item: DimItem, e?: MouseEvent, equip?: boolean) => void
+  add: (item: DimItem, e?: MouseEvent, equip?: boolean) => void,
+  onApplySocketOverrides: (item: DimItem, socketOverrides: SocketOverrides) => void
 ) {
   if (!loadout) {
     return;
@@ -217,6 +322,9 @@ function fillLoadoutFromEquipped(
       !itemsByBucket[item.bucket.hash].some((i) => i.equipped)
     ) {
       add(item, undefined, true);
+      if (item.bucket.hash === BucketHashes.Subclass) {
+        createSocketOverridesFromEquipped(item, onApplySocketOverrides);
+      }
     } else {
       infoLog('loadout', 'Skipping', item, { itemsByBucket, bucketId: item.bucket.hash });
     }